Large scale rewrite to decouple and improve components; Introduced first-class annotation support
All publicly exposed classes are now represented as abstract interfaces,
deferring functionality to concrete internal implementations. This
should make the whole reflection system more maintainable in the future.

Through sheer force of will, annotations/attributes have been brought to
Turing! To start using them, import the "annotations" module and create
an empty procedure with the word "annotation" before it. Now when you
call that procedure in front of a code construct, it and all its
arguments are made available as annotation data to reflective objects.

The Opcodes class has had all its exported opcodes marked as being
"not-dot", so code that was previously Opcodes.FOO can now be written as
FOO.